# [C#] Optimize reading in PushBackReader by removing caching into nullable char.