I've found that PushBackReader is storing next character in nullable char field. And accessing it is visible in CPU profiler:
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/081b488f-442e-47e4-af15-198556f467e9)

This PR is converting it to int and checks for -1 instead of .HasValue.
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/05652564-cc67-4b9f-83d7-d8a05f72f2dc)

Using [benchmark](https://github.com/protocolbuffers/protobuf/pull/15795#issuecomment-1980222331) gives following results
Main:
// * Summary *

BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3235/23H2/2023Update/SunValley3)
11th Gen Intel Core i9-11900K 3.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.200
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI

Job=DefaultJob
| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     295.7 ns |   1.01 KB |
| Big    | 265,443.3 ns |  642.7 KB |

PR
| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     241.6 ns |   1.01 KB |
| Big    | 207,593.5 ns |  642.7 KB |