@jskeet 2024-03-07
I'll have a closer look, but this really feels like it's hurting the readability of the code.

I don't know when I'll have time to look at any of this though. I don't think there's very much point in creating an ever-growing backlog of micro-optimisations for JSON parsing, to be honest.