@TrayanZapryanov 2024-03-07
I understand that JSON parsing is not in the focus and already start a change in our code to migrate to binary serialization.
But still see a value in all these PRs as they slowly closing the gap between both kinds of deserializations.
I haven't saw somewhere a guideline that is advising people to prefer one over the other and maybe this is what is missing and should be documented. Please fill free in closing all PRs which you don't like - I am really satisfied that found/suggested them and once combined they lead to following results:
Start:
| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     295.7 ns |   1.01 KB |
| Big    | 265,443.3 ns |  642.7 KB |

All in one:
| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     261.2 ns |     456 B |
| Big    | 190,585.2 ns |  118008 B |

Thanks for review and guidance.