@jskeet 2024-03-07
It's not that there won't eventually be a benefit - it's that I don't see much point in stacking up an ever-growing number of them which will potentially end up causing merge conflicts with one another. I would suggest staggering your effort on this - work out which *one* PR you most want to land next, I can try to find time to review that (but it'll take a while - Protobuf isn't my main project, and I have other Protobuf work to do as well) and get it merged, then we can move on to the next one.