@jskeet 2024-03-15
As ever, I'll have a look when I get time, but there's only so much time I can put into these PRs. (I'm expecting to basically put time into one per week.)