@jskeet 2024-03-20
I've had a closer look, and given that it's purely internal within the class (and nested classes) *and* that the performance difference is significant, I think it's worth going with the `int` version.

I've created an alternative for the PR [here](https://github.com/protocolbuffers/protobuf/commit/454abe07ce1250767b80777309a3f1c124e4684e) - please have a look and see if that looks okay to you. If it does, I'll push it through internally like the last PR.

There are a few changes in there compared with your commit, to take further advantage of just having a single value - as well as more comments to make up for the "odd" choice of type.